# Optimized heap snapshot write