As discussed on #47134, most the time of taking a heap snapshot is spent writing the result. Just combining the `ios_printf` calls makes that significantly faster. Also calling `ios_putc()` instead of `ios_printf()` takes off another half second or so.

```julia
using Profile;
@time Profile.take_heap_snapshot(tempname());

# Before:
# apaz[x][master][~/git/julia]$ ./julia snap.jl 
#  5.397448 seconds (20 allocations: 1.930 KiB, 34.34% gc time)

# After:
# apaz[x][ap/snapshot-write][~/git/julia]$ ./julia snap.jl 
#  3.808458 seconds (20 allocations: 1.930 KiB, 47.86% gc time)
```

```julia
using LoopVectorization;
using Profile;
@time Profile.take_heap_snapshot(tempname());

# Before:
# apaz[x][master][~/git/julia]$ ./julia biggersnap.jl 
#  7.153016 seconds (20 allocations: 1.930 KiB, 35.07% gc time)

# After:
# apaz[x][ap/snapshot-write][~/git/julia]$ ./julia biggersnap.jl
#  5.280476 seconds (20 allocations: 1.930 KiB, 49.39% gc time)
```

These "benchmarks" are not very scientific. This is a single run on my laptop, while using it for other things, and there's significant (~.5s) variability between runs. Still, a noteable improvement.